Added FPGA support!
Lots and lots of changes, adding support for FPGA targets. The heart of it is the new
conditionally-compiled NOR gate models, which need a 50MHz clock for timing. There's
also lots of changes to the generation scripts in support of generating synthesizable
code. In summary, the FPGA build looks like this::
* Open drain buses are implemented using wands instead of tri-state + pullups
* All chip-level modules must reside in the same file due to altera's handling of wand
  (it can't cross module boundaries)
* Proxy wires are used to connect component modules to wand buses; this prevents module crossing
* Components and parts have been broken into separate files to appease Quartus
* '_NC' wires are gone, now using implicit no connects
* Internal net names include the module number for clarification/deconfliction

For the most part, the Icarus target remains the same, aside from using the same
set of new files for components/parts. I also spruced up the I/O logic in the
backplane generator, so signals that make it onto the backplane that have no
internal destination are marked as outputs. This gets monitor signals automatically
routed out of the top-level module, which makes connecting them to pins easy.

Quartus project that instantiates this to follow.