Added a memory module that implements fixed memory in the form of flash.
SST39VF200A is a simple model of the flash chip I'll be using,
and while it needs a bit more work it operates correctly.
I added a new script, bin_to_verilog_rom.py, that converts
a binary file to a ROM image for inclusion in the flash driver.
Theoretically programs should synthesize to small enough to
fit in an FPGA using the switch statement implementation of flash.
FPGA main file has not yet been updated.