Allow the flash module to use serial memory chips.
The previous massive switch approach, while technically synthesizable,
was going to take >12 hours to actually synthesize, so I had to find a
different approach to storing FSW. The DE0-Nano doesn't have quite
enough memory in its memory cells to fit everything (plus those will be
more useful for RAM anyways). The only other option was to stick the FSW
in the EPCS64, the FPGA's configuration device.

Transaction times with the EPCS are somewhat slow -- certainly slower
than the duration of the SBF pulse. To give the transaction enough time
to finish, I had to redo the rope simulation logic a little to determine
and latch the target address in advance of SBF.

The FPGA-targeted flash module now listens to its chip enable pin, and
when enabled, starts and EPCS read at the currently selected address.
It then holds the read word until it's deselected.

The codegen scripts needed several changes (and a KiCAD patch) to handle
the idea of simulated 'chips' needing to have extra I/O to talk to FPGA
peripherals. There's now FPGA#inputs and FPGA#outputs component fields
that the scripts look for, and introduce relevant nets when they're
present. I've also started treating tristate-facing nets as 'wor's
rather than 'wand's, since that's closer to how the sense amplifier
lines actually work (this will be needed for the addition of erasable
memory).