Added FPGA#bidir codegen flag support for bidirectional pins.
I haven't yet found a way to get bidir ports to work the way I need them
to for the FPGA, so instead I'm taking the easy way out and duplicating
the ports for pins that have this new flag. Used in conjunction with the
FPGA#OD codegen flag, outputs are still proxied, and the duplicated
inputs are taken from the original signals, rather than the proxies.