Initialize all of the RAM to have the right parity
Magnetic core memory is nonvolatile, and the AGC developers took full
advantage of that fact. Colossus does a lot of accesses to memory that
has not been initialized. However, any access to memory without correct
parity causes a GOJAM. This process does manage to set the right parity
on the word being accessed, so it will succeed the next attempt. It does
mean that the FSW will reboot many, many times when run with fresh RAM,
making it one erasable memory access further each time.

It's much easier to just set the words to have right parity in
simulation.