Switch to fast read for DE0-Nano fixed memory
The alarms module was freaking out about tons of parity errors. It turns
out this is because I failed to account for register clearing when
calculating out the required deadline for the transaction to be done --
the word must be read before CGG falls, rather than before SBF ends. By
random luck, the rising edge of the memory clock that latched in the
parity bit happened within some nanoseconds of the falling edge of CGG,
which led to effectively a random parity. To fix this, I've switched
from using the regular read operation to fast read. While this adds an
additional byte of overhead to the transaction, it lets me double the
memory clock speed, which leaves us with plenty of time to spare.