Fixed the instruction decoder for long traces.
It was, of course, not a GTKWave bug. GTKWave uses pipes to communicate
with its subprocess, and apparently decoding everything and printing on
the fly is too slow, and eventually causes a backup and subsequent hang.
Everything seems to go much more smoothly if I just buffer up all the
lines before doing any decoding.