JTAG Monitor: added instruction step support, and read access to most registers.
I'm not sure how I'm supposed to know what X is, so I'm wondering if the
Monitor access to Y and U are strictly for insight into STORE and FETCH.
Either way, I've reserved some registers for them all. The instruction
register width had to increase to 5 bits to fit all the registers, and
accomodate for the breakpoint stuff that is coming up next.